Improved waypoints, documentation and added text panel to video relics
Improved waypoints material for better visibility in Forest scene. Also added more screenshots from the scene. Included UI Panel at the start of the experience with a short one-liner to guide your users on how to navigate the environment. Included  text panels and placed it below the stations highlighting the topic for that station.